server: use routable url for user creation
Our httprouter cleans up routes that appear to be missing a trailing
slash by redirecting clients. That means that the url `*/users` cannot
be hit, since there is a live `*/users/` url - the router will always
redirect the url without the slash. This pull request  attempts to avoid
the issue by moving the endpoint.