encode pw from client before further processing
The changes done in #304 makes dex expect that clients send base64
encoded passwords. I believe that this is an assumption we cannot
make.

Changed the server so it base64 encodes the password from the client
before further processing internally in Dex.

Fixed the tests so they create a mockServer with a base64-encoded
password repo and a mockClient that transmits a plain password.

Partially fixes #335