storage: allow to store email_verified flag for local passwords
email_verified claim is attached by Dex to every token. For local
passwords it's always true, and it creates problems for apps that rely
on that kind of information coming from ID tokens.

Store email_verified flag in the storage and allow it to be updated
using UpdatePassword.

The default value is set to 'true' for compatibility with clients
migrating from previous Dex versions.

Signed-off-by: Pavel Borzenkov <pavel.borzenkov@gmail.com>