Added User Info endpoint
1. The AccessToken is filled with Claims, Connector Data and ConnectorID from the AuthCode.
2. It is then Serialized and sent to the user
3. The user may, then, do a request to the /userinfo endpoint using this token.

By default, when dex gets a /userinfo request, the token is deserialized, the Claims originally copied from AuthCode are extracted from it and sent back to the user as the User Info response payload.

This should work for all the connectors and was tested with the LDAP and OIDC connectors.

If the connector used to authenticate the user implements the GetUserInfo interface, then the responsibility of generating the User Info payload is delegated to it.

This is the case of the Oidc connector.
It uses the real AccessToken, previously stored as a field in the ConnectorData, to forward a genuine /userinfo request to the original IdP.
The Claims extracted from the AuthCode are used to generate the initial user info payload object, and this object is extended with whatever claims are returned by the original IdP.