connector/authproxy: supports groups claim
The groups are collected from the X-Remote-Group headers as it is
proposed at https://kubernetes.io/docs/admin/authentication/#authenticating-proxy.

The X-Remote-User header is handled as email address only if it contains
"@" character, and the Username is derived before the "@" sign.

The UserID of the Identity is not set because we do not have any
knowledge of the proxied authentication how should that be handled.