Only replace whole config-keys with environments
Reflect over and walk the loaded configuration and replace any $ENV keys
with their corresponding environment variable, but only of the whole key
matches.

This makes it possible to include '$' in configuration keys (e.g.
configured passwords; see #1619). It also removes the possibility for
injecting further JSON/yaml in the middle of a document through
env-expansion (see #1099).

On the flip-side, this makes it impossible to do env-expansion on
anything non-string (i.e. numbers) through environment variables, as
`"$name"` is an invalid number.

Previously, this was done by running the JSON through os.ExpandEnv(),
which does not do escaping, making it impossible to embed a '$' anywhere
in the storage or connector configurations.