handlePasswordGrant: insert connectorData into OfflineSession
Using dex with the kubernetes oidc client results in the creation
of an Offline Session with empty ConnectorData.

This change will insert the ConnectorData from the initial Login
into the OfflineSession, as already done in `handlePasswordLogin`.

Not doing this results in failures to refresh a token.
I've observed this with the ldap connetor, other connectors might
be affected as well.

The ldap connector will fail with this error message:
`failed to refresh identity: ldap: failed to unmarshal internal data: unexpected end of JSON input`

The `Refresh` method of the ldap connector will first try to unmarshal
the ConnectorData, which was empty, as dex wouldn't
store it in storage after a successful login.

Signed-off-by: Henning Surmeier <h.surmeier@mittwald.de>