Add support for group and email headers to authproxy connector
This commit adds several new options to the authproxy connector to allow additional user
attributes to be extracted.

A new emailHeader config option on the connector allows different values for the UserID
Email fields of the identity to be used, if the option is not specified then Email will
be populated from UserID as before.

User groups may also be extracted from headers in one of two different ways.
In order to extract group information the groupHeader config option must be specified.
If the groupSeparator option is specified and non-empty then the group header is treated
as a delimited list of group names, for example given groupSeparator = "," then the
expected format of header would be

X-Remote-Group: allUsers,admins

Alternatively if groupSeparator is not provided then the header is treated as a repeated
list of group names, for example

X-Remote-Group: allUsers
X-Remote-Group: admins

Note that the connector behaviour remains unchanged in all cases where the new config
options have not been used, ensuring backwards compatibility for existing installations.

Signed-off-by: GCHQDeveloper73046 <92861835+GCHQDeveloper73046@users.noreply.github.com>