Add ForceQueryResponseMode to OIDC connector
Adds a ForceQueryResponseMode configuration option which can be used to
enforce the response_mode parameter set to `query` in the LoginURL and
is set to false by default.

Although not `query` is the default response_mode according to the OIDC
specification some implementations use other response_modes by default
and therefore require this option to be set to work with Dex.

Fixes #2580

Signed-off-by: Johannes Tandler <johannes.tandler@zeiss.com>