Support multiple LDAP hosts
Often there is a need to iterate through multiple LDAP hosts in the
event that one has failed. DNS is one possible solution, but there
are cases where that may not be desired. This change allows users to
specify a comma-separated list of LDAP hosts to iterate through in case
one is unreachable.

Signed-off-by: Ryan Canty <jrcanty@gmail.com>