microsoft: get rid of User.Read scope which leads to additional consents
It seems that when a user is already logged in with Microsoft in a
browser and dex requires User.Read scope it forces Microsoft to show
an additional screen from time to time:

-----------
Permissions Requested
...
This app would like to:
Sign in and read user profile
-----------

If instead a more standart OIDC connector procedure is used with
openid,profile,email scopes - it doesn't happen and no microsoft
screens are shown.

This patch kills the need for User.Read scope and follows a more
standart OIDC connector procedure for Microsoft login.
Note, microsoft doesn't fully follow spec, so issuer field in well-known
configuration and token is verified manually.

Signed-off-by: Alex Lyashko <alex.lyashko@gmail.com>