fix: remove redundant scope check from Google connector
Since `groups` are not considered a valid scope by Google, we cannot
pass to the `scope`. Although @damieva introduced `claimMapping` in
https://github.com/dexidp/dex/issues/2653 to address this issue, it does
not seem to work with the current implementation of the Google
connector.

This commit resolves the issue by removing the unnecessary `group` scope
check, as `adminSrv` is only initialized when the user wants to retrieve
groups.

Signed-off-by: Sunghoon Kang <hoon@akuity.io>