Do not escape password for LDAP connectors
With the change introduced in https://github.com/dexidp/dex/pull/3372 Dex declines passwords that contain special characters. Since password is not passed to any kind of filters, it is safe to pass a password as is. No LDAP query injections are possible.

This commit is a revert of password escaping.

Signed-off-by: m.nabokikh <maksim.nabokikh@flant.com>