google: add support for groups access without service account
The Google connector has support for retrieving groups the user email
is a member of using the Admin SDK API. This API requires using an
authenticated client sourced from a service account granted domain-wide
delegation. This account would be highly-privileged and give access to all
Workspace users, which is less than ideal.

This change adds an alternative method to retrieve group membership
using the Cloud Identity API. When using the SearchDirectGroups API
endpoint, we can provide an access token that has been authorized with
the appropriate Cloud Identity API scope and search for transitive
membership.

In contrast to the existing method where the user is expected to specify
the "groups" scope, this one relies on ther user setting one of the
required Cloud Identity scope explicitly:
- https://www.googleapis.com/auth/cloud-platform
- https://www.googleapis.com/auth/cloud-identity.groups
- https://www.googleapis.com/auth/cloud-identity.groups.readonly

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>